Try to disable the magisk notification and debug log for nethunter app per installation.
By default the Magisk will log and pop up a toast message on screen every time you execute a runtime with root.
It will be very annoying to keep show you a bunch of toast message, so it will try to disable it by modifying the magisk db file using sqlite3 tool.